split the common of contract
